# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Yet Another Node.js Blogg Application API",
    "description": "Yet Another Node.js Blogg Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "tags": [
    {
      "name": "Users",
      "description": "API for users in the system"
    },
    {
      "name": "Waterings",
      "description": "API for waterings in the system"
    },
    {
      "name": "Watering management",
      "description": "API for watering management in the system"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Create new user in system",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get all users in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
    "/users/{userId}": {
      "parameters": [
        {
          "name": "userId",
          "in": "path",
          "required": true,
          "description": "ID of user that we want to find",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user with given ID",
        "responses": {
          "200": {
            "description": "User is found",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete user with given ID",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "User is deleted",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "put": {
        "summary": "Update user with give ID",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User with new values of properties",
            "schema": {
              "$ref": "#/definitions/updateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User is updated",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/waterings": {
      "post": {
        "tags": [
          "Waterings"
        ],
        "summary": "Creat new watering in system",
        "parameters": [
          {
            "name": "watering",
            "in": "body",
            "description": "Watering that we want to create",
            "schema": {
              "$ref": "#/definitions/Watering"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New watering is created",
            "schema": {
              "$ref": "#/definitions/Watering"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Waterings"
        ],
        "summary": "Get all waterings in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Waterings"
            }
          }
        }
      }
    },
    "/waterings/{wateringId}": {
      "parameters": [
        {
          "name": "wateringId",
          "in": "path",
          "required": true,
          "description": "ID of watering that we want to find",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Waterings"
        ],
        "summary": "Get watering with given ID",
        "responses": {
          "200": {
            "description": "Watering is found",
            "schema": {
              "$ref": "#/definitions/Watering"
            }
          }
        }
      }
    },
    "/wateringManagement/power": {
      "post": {
        "tags": [
          "Watering management"
        ],
        "summary": "Start (true) or end (false) a watering in system",
        "parameters": [
          {
            "name": "wateringPower",
            "in": "body",
            "description": "Start (true) or end (false) a watering",
            "schema": {
              "$ref": "#/definitions/wateringPower"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Power of watering is managed",
            "schema": {
              "$ref": "#/definitions/wateringPower"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "required": [
        "email"
      ],
      "properties": {
        "email": {
          "type": "string",
          "uniqueItems": true
        },
        "lastName": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        }
      }
    },
    "updateUser": {
      "required": [
        "email",
        "_id"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "uniqueItems": true
        },
        "email": {
          "type": "string",
          "uniqueItems": true
        },
        "lastName": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        }
      }
    },
    "Users": {
      "type": "array",
      "$ref": "#/definitions/User"
    },
    "Watering": {
      "required": [
        "id_user",
        "name"
      ],
      "properties": {
        "id_user": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "Waterings": {
      "type": "array",
      "$ref": "#/definitions/Watering"
    },
    "wateringPower": {
      "required": [
        "id_user",
        "power"
      ],
      "properties": {
        "id_watering": {
          "type": "string"
        },
        "power": {
          "type": "boolean"
        },
        "time": {
          "type": "string",
          "format": "date-time"
        }
      }
    }
  }
}